hore

+++ si4alaravel/app/Http/Controllers/FakultasController.php
@@ -32,9 +32,10 @@ class FakultasController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {
+    {  if ($request->user()->cannot('create', Fakultas::class)) { abort(403); }
         //
         //validasi input
+        
         $input = $request->validate(
             [
                 'nama'=>'required|unique:fakultas',
@@ -75,8 +76,11 @@ class FakultasController extends Controller
      */
     public function update(Request $request, $fakultas)
     {
+            $fakultas = Fakultas::findorfail($fakultas);
+
+        if ($request->user()->cannot('update', $fakultas)) { abort(403); }
         
-        $fakultas = Fakultas::findorfail($fakultas);
+    
         //validasi input
         $input = $request->validate(
             [
@@ -97,11 +101,13 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         //
         $fakultas=Fakultas::findOrFail($fakultas);
 
+        if ($request->user()->cannot('update', $fakultas)) { abort(403); }
+        
         $fakultas->delete();
 
         return redirect()->route('fakultas.index')->with('success', 'Fakultas berhasil dihapus');
